Stabilize KDF APIs
