+++ aws-lc-rs/tests/kdf_test.rs
@@ -5,14 +5,12 @@
 
 use std::error::Error;
 
-use aws_lc_rs::{
-    test, test_file,
-    unstable::kdf::{
-        get_kbkdf_ctr_hmac_algorithm, get_sskdf_digest_algorithm, get_sskdf_hmac_algorithm,
-        kbkdf_ctr_hmac, sskdf_digest, sskdf_hmac, KbkdfCtrHmacAlgorithm, KbkdfCtrHmacAlgorithmId,
-        SskdfDigestAlgorithm, SskdfDigestAlgorithmId, SskdfHmacAlgorithm, SskdfHmacAlgorithmId,
-    },
+use aws_lc_rs::kdf::{
+    get_kbkdf_ctr_hmac_algorithm, get_sskdf_digest_algorithm, get_sskdf_hmac_algorithm,
+    kbkdf_ctr_hmac, sskdf_digest, sskdf_hmac, KbkdfCtrHmacAlgorithm, KbkdfCtrHmacAlgorithmId,
+    SskdfDigestAlgorithm, SskdfDigestAlgorithmId, SskdfHmacAlgorithm, SskdfHmacAlgorithmId,
 };
+use aws_lc_rs::{test, test_file};
 
 #[derive(Clone, Copy)]
 enum SskdfVariant {
